+++ govcd/access_control_catalog_test.go
@@ -51,8 +51,8 @@ func (vcd *TestVCD) Test_AdminCatalogAccessControl(check *C) {
 
 	orgInfo, err := adminCatalog.getOrgInfo()
 	check.Assert(err, IsNil)
-	check.Assert(orgInfo.id, Equals, extractUuid(adminorg.AdminOrg.ID))
-	check.Assert(orgInfo.name, Equals, adminorg.AdminOrg.Name)
+	check.Assert(orgInfo.OrgId, Equals, extractUuid(adminorg.AdminOrg.ID))
+	check.Assert(orgInfo.OrgName, Equals, adminorg.AdminOrg.Name)
 
 	err = adminCatalog.Delete(true, true)
 	check.Assert(err, IsNil)
@@ -82,8 +82,8 @@ func (vcd *TestVCD) Test_CatalogAccessControl(check *C) {
 
 	orgInfo, err := catalog.getOrgInfo()
 	check.Assert(err, IsNil)
-	check.Assert(orgInfo.id, Equals, extractUuid(adminorg.AdminOrg.ID))
-	check.Assert(orgInfo.name, Equals, adminorg.AdminOrg.Name)
+	check.Assert(orgInfo.OrgId, Equals, extractUuid(adminorg.AdminOrg.ID))
+	check.Assert(orgInfo.OrgName, Equals, adminorg.AdminOrg.Name)
 
 	err = catalog.Delete(true, true)
 	check.Assert(err, IsNil)