+++ govcd/nsxt_application_profile_test.go
@@ -176,7 +176,7 @@ func getResultCountByScope(scope string, check *C, vcd *TestVCD) int {
 		Resulttotal int `json:"resultTotal"`
 	}{}
 
-	err = vcd.vdc.client.OpenApiGetItem(apiVersion, urlRef, queryParams, &result)
+	err = vcd.vdc.client.OpenApiGetItem(apiVersion, urlRef, queryParams, &result, nil)
 	check.Assert(err, IsNil)
 	return result.Resulttotal
 }